とりあえず配列で動く
