+++ .github/workflows/ci.yml
@@ -30,5 +30,5 @@ jobs:
           chmod +x test/simple-test.sh
           ./test/simple-test.sh
 
-      - name: Valgrind check
-        run: valgrind --leak-check=full ./hotrace || true
+      # - name: Valgrind check
+      #   run: valgrind --leak-check=full ./hotrace || true
