+++ srcs/_utils/utils.h
@@ -11,13 +11,13 @@
 /* ************************************************************************** */
 
 #ifndef UTILS_H
-#define UTILS_H
+# define UTILS_H
 
-#include <unistd.h>
-#include <stdlib.h>
+# include <stdlib.h>
+# include <unistd.h>
 
-int ft_strlen(const char *s);
-int ft_strcmp(const char *s1, const char *s2);
-char *ft_strdup(const char *s);
+int		ft_strlen(const char *s);
+int		ft_strcmp(const char *s1, const char *s2);
+char	*ft_strdup(const char *s);
 
 #endif
