+++ srcs/v1-map/search-map.c
@@ -1,31 +1,11 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   store-array.c                                      :+:      :+:    :+:   */
+/*   search-map.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: yui <yui@student.42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/06/28 00:01:21 by yui               #+#    #+#             */
-/*   Updated: 2025/06/28 13:41:47 by yui              ###   ########.fr       */
+/*   Created: 2025/06/28 00:00:03 by yui               #+#    #+#             */
+/*   Updated: 2025/06/28 13:35:46 by yui              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
-
-#include <utils.h>
-
-int	store(const char *key, const char *value, void *ctx)
-{
-	const char	***array = (const char ***)ctx;
-	static int	i = 0;
-
-	if (!key || !value || !ctx)
-		return (-1);
-	array[i][0] = ft_strdup(key);
-	if (!array[i][0])
-		return (-1);
-	array[i][1] = ft_strdup(value);
-	if (!array[i][1])
-		return (-1);
-	if (++i >= 1024)
-		return (-1);
-	return (0);
-}
